+++ CHANGES.md
@@ -5,6 +5,7 @@
 ## Changed
 
 + relm4: Pass model in connect_components function of the Widgets trait
++ relm4: Added with_app method that allows passing an existing gtk::Appliaction to Relm4
 + book: Many book improvements by [tronta](https://github.com/tronta)
 
 ## Added